+++ processing/src/main/java/org/apache/druid/query/filter/InDimFilter.java
@@ -26,6 +26,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
+import com.google.common.base.Predicate;
 import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 import com.google.common.collect.ImmutableSet;
@@ -37,17 +38,29 @@ import com.google.common.collect.TreeRangeSet;
 import com.google.common.hash.Hasher;
 import com.google.common.hash.Hashing;
 import it.unimi.dsi.fastutil.ints.IntArrayList;
+import it.unimi.dsi.fastutil.ints.IntIterable;
+import it.unimi.dsi.fastutil.ints.IntIterator;
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import it.unimi.dsi.fastutil.longs.LongArrayList;
 import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
+import org.apache.druid.collections.bitmap.ImmutableBitmap;
 import org.apache.druid.common.config.NullHandling;
+import org.apache.druid.java.util.common.IAE;
 import org.apache.druid.java.util.common.StringUtils;
+import org.apache.druid.query.BitmapResultFactory;
 import org.apache.druid.query.cache.CacheKeyBuilder;
 import org.apache.druid.query.extraction.ExtractionFn;
+import org.apache.druid.query.filter.vector.VectorValueMatcher;
+import org.apache.druid.query.filter.vector.VectorValueMatcherColumnProcessorFactory;
 import org.apache.druid.query.lookup.LookupExtractionFn;
 import org.apache.druid.query.lookup.LookupExtractor;
+import org.apache.druid.segment.ColumnSelector;
+import org.apache.druid.segment.ColumnSelectorFactory;
 import org.apache.druid.segment.DimensionHandlerUtils;
-import org.apache.druid.segment.filter.InFilter;
+import org.apache.druid.segment.IntIteratorUtils;
+import org.apache.druid.segment.column.BitmapIndex;
+import org.apache.druid.segment.filter.Filters;
+import org.apache.druid.segment.vector.VectorColumnSelectorFactory;
 
 import javax.annotation.Nullable;
 import java.nio.charset.StandardCharsets;
@@ -56,12 +69,14 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class InDimFilter implements DimFilter
+public class InDimFilter implements DimFilter, Filter
 {
   // determined through benchmark that binary search on long[] is faster than HashSet until ~16 elements
   // Hashing threshold is not applied to String for now, String still uses ImmutableSortedSet
@@ -74,12 +89,10 @@ public class InDimFilter implements DimFilter
   private final ExtractionFn extractionFn;
   @Nullable
   private final FilterTuning filterTuning;
-  private final Supplier<DruidLongPredicate> longPredicateSupplier;
-  private final Supplier<DruidFloatPredicate> floatPredicateSupplier;
-  private final Supplier<DruidDoublePredicate> doublePredicateSupplier;
+  private final DruidPredicateFactory predicateFactory;
 
   @JsonIgnore
-  private byte[] cacheKey;
+  private final Supplier<byte[]> cacheKeySupplier;
 
   @JsonCreator
   public InDimFilter(
@@ -91,8 +104,42 @@ public class InDimFilter implements DimFilter
       @JsonProperty("filterTuning") @Nullable FilterTuning filterTuning
   )
   {
-    Preconditions.checkNotNull(dimension, "dimension can not be null");
-    Preconditions.checkArgument(values != null, "values can not be null");
+    this(
+        dimension,
+        values,
+        extractionFn,
+        filterTuning,
+        null
+    );
+  }
+
+  /**
+   * This constructor should be called only in unit tests since accepting a Collection makes copying more likely.
+   */
+  @VisibleForTesting
+  public InDimFilter(String dimension, Collection<String> values, @Nullable ExtractionFn extractionFn)
+  {
+    this(
+        dimension,
+        values instanceof Set ? (Set<String>) values : new HashSet<>(values),
+        extractionFn,
+        null,
+        null
+    );
+  }
+
+  /**
+   * Internal constructor.
+   */
+  private InDimFilter(
+      final String dimension,
+      final Set<String> values,
+      @Nullable final ExtractionFn extractionFn,
+      @Nullable final FilterTuning filterTuning,
+      @Nullable final DruidPredicateFactory predicateFactory
+  )
+  {
+    Preconditions.checkNotNull(values, "values cannot be null");
 
     // The values set can be huge. Try to avoid copying the set if possible.
     // Note that we may still need to copy values to a list for caching. See getCacheKey().
@@ -101,21 +148,18 @@ public class InDimFilter implements DimFilter
     } else {
       this.values = values.stream().map(NullHandling::emptyToNullIfNeeded).collect(Collectors.toSet());
     }
-    this.dimension = dimension;
+
+    this.dimension = Preconditions.checkNotNull(dimension, "dimension cannot be null");
     this.extractionFn = extractionFn;
     this.filterTuning = filterTuning;
-    this.longPredicateSupplier = getLongPredicateSupplier();
-    this.floatPredicateSupplier = getFloatPredicateSupplier();
-    this.doublePredicateSupplier = getDoublePredicateSupplier();
-  }
 
-  /**
-   * This constructor should be called only in unit tests since it creates a new hash set wrapping the given values.
-   */
-  @VisibleForTesting
-  public InDimFilter(String dimension, Collection<String> values, @Nullable ExtractionFn extractionFn)
-  {
-    this(dimension, new HashSet<>(values), extractionFn, null);
+    if (predicateFactory != null) {
+      this.predicateFactory = predicateFactory;
+    } else {
+      this.predicateFactory = new InFilterDruidPredicateFactory(extractionFn, this.values);
+    }
+
+    this.cacheKeySupplier = Suppliers.memoize(this::computeCacheKey);
   }
 
   @JsonProperty
@@ -149,26 +193,7 @@ public class InDimFilter implements DimFilter
   @Override
   public byte[] getCacheKey()
   {
-    if (cacheKey == null) {
-      final List<String> sortedValues = new ArrayList<>(values);
-      sortedValues.sort(Comparator.nullsFirst(Ordering.natural()));
-      final Hasher hasher = Hashing.sha256().newHasher();
-      for (String v : sortedValues) {
-        if (v == null) {
-          hasher.putInt(0);
-        } else {
-          hasher.putString(v, StandardCharsets.UTF_8);
-        }
-      }
-      cacheKey = new CacheKeyBuilder(DimFilterUtils.IN_CACHE_ID)
-          .appendString(dimension)
-          .appendByte(DimFilterUtils.STRING_SEPARATOR)
-          .appendByteArray(extractionFn == null ? new byte[0] : extractionFn.getCacheKey())
-          .appendByte(DimFilterUtils.STRING_SEPARATOR)
-          .appendByteArray(hasher.hash().asBytes())
-          .build();
-    }
-    return cacheKey;
+    return cacheKeySupplier.get();
   }
 
   @Override
@@ -188,6 +213,197 @@ public class InDimFilter implements DimFilter
       );
     }
     return inFilter;
+  }
+
+  @Override
+  public Filter toFilter()
+  {
+    return this;
+  }
+
+  @Nullable
+  @Override
+  public RangeSet<String> getDimensionRangeSet(String dimension)
+  {
+    if (!Objects.equals(getDimension(), dimension) || getExtractionFn() != null) {
+      return null;
+    }
+    RangeSet<String> retSet = TreeRangeSet.create();
+    for (String value : values) {
+      String valueEquivalent = NullHandling.nullToEmptyIfNeeded(value);
+      if (valueEquivalent == null) {
+        // Case when SQL compatible null handling is enabled
+        // Range.singleton(null) is invalid, so use the fact that
+        // only null values are less than empty string.
+        retSet.add(Range.lessThan(""));
+      } else {
+        retSet.add(Range.singleton(valueEquivalent));
+      }
+    }
+    return retSet;
+  }
+
+  @Override
+  public Set<String> getRequiredColumns()
+  {
+    return ImmutableSet.of(dimension);
+  }
+
+  @Override
+  public <T> T getBitmapResult(BitmapIndexSelector selector, BitmapResultFactory<T> bitmapResultFactory)
+  {
+    if (extractionFn == null) {
+      final BitmapIndex bitmapIndex = selector.getBitmapIndex(dimension);
+      return bitmapResultFactory.unionDimensionValueBitmaps(getBitmapIterable(values, bitmapIndex));
+    } else {
+      return Filters.matchPredicate(
+          dimension,
+          selector,
+          bitmapResultFactory,
+          predicateFactory.makeStringPredicate()
+      );
+    }
+  }
+
+  @Override
+  public double estimateSelectivity(BitmapIndexSelector indexSelector)
+  {
+    if (extractionFn == null) {
+      final BitmapIndex bitmapIndex = indexSelector.getBitmapIndex(dimension);
+      return Filters.estimateSelectivity(
+          bitmapIndex,
+          IntIteratorUtils.toIntList(getBitmapIndexIterable(values, bitmapIndex).iterator()),
+          indexSelector.getNumRows()
+      );
+    } else {
+      return Filters.estimateSelectivity(
+          dimension,
+          indexSelector,
+          predicateFactory.makeStringPredicate()
+      );
+    }
+  }
+
+  @Override
+  public ValueMatcher makeMatcher(ColumnSelectorFactory factory)
+  {
+    return Filters.makeValueMatcher(factory, dimension, predicateFactory);
+  }
+
+  @Override
+  public VectorValueMatcher makeVectorMatcher(final VectorColumnSelectorFactory factory)
+  {
+    return DimensionHandlerUtils.makeVectorProcessor(
+        dimension,
+        VectorValueMatcherColumnProcessorFactory.instance(),
+        factory
+    ).makeMatcher(predicateFactory);
+  }
+
+  @Override
+  public boolean canVectorizeMatcher()
+  {
+    return true;
+  }
+
+  @Override
+  public boolean supportsRequiredColumnRewrite()
+  {
+    return true;
+  }
+
+  @Override
+  public Filter rewriteRequiredColumns(Map<String, String> columnRewrites)
+  {
+    String rewriteDimensionTo = columnRewrites.get(dimension);
+    if (rewriteDimensionTo == null) {
+      throw new IAE("Received a non-applicable rewrite: %s, filter's dimension: %s", columnRewrites, dimension);
+    }
+
+    if (rewriteDimensionTo.equals(dimension)) {
+      return this;
+    } else {
+      return new InDimFilter(
+          rewriteDimensionTo,
+          values,
+          extractionFn,
+          filterTuning,
+          predicateFactory
+      );
+    }
+  }
+
+  @Override
+  public boolean supportsBitmapIndex(BitmapIndexSelector selector)
+  {
+    return selector.getBitmapIndex(dimension) != null;
+  }
+
+  @Override
+  public boolean shouldUseBitmapIndex(BitmapIndexSelector selector)
+  {
+    return Filters.shouldUseBitmapIndex(this, selector, filterTuning);
+  }
+
+  @Override
+  public boolean supportsSelectivityEstimation(ColumnSelector columnSelector, BitmapIndexSelector indexSelector)
+  {
+    return Filters.supportsSelectivityEstimation(this, dimension, columnSelector, indexSelector);
+  }
+
+  @Override
+  public String toString()
+  {
+    final DimFilterToStringBuilder builder = new DimFilterToStringBuilder();
+    return builder.appendDimension(dimension, extractionFn)
+                  .append(" IN (")
+                  .append(Joiner.on(", ").join(Iterables.transform(values, StringUtils::nullToEmptyNonDruidDataString)))
+                  .append(")")
+                  .appendFilterTuning(filterTuning)
+                  .build();
+  }
+
+  @Override
+  public boolean equals(Object o)
+  {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    InDimFilter that = (InDimFilter) o;
+    return values.equals(that.values) &&
+           dimension.equals(that.dimension) &&
+           Objects.equals(extractionFn, that.extractionFn) &&
+           Objects.equals(filterTuning, that.filterTuning);
+  }
+
+  @Override
+  public int hashCode()
+  {
+    return Objects.hash(values, dimension, extractionFn, filterTuning);
+  }
+
+  private byte[] computeCacheKey()
+  {
+    final List<String> sortedValues = new ArrayList<>(values);
+    sortedValues.sort(Comparator.nullsFirst(Ordering.natural()));
+    final Hasher hasher = Hashing.sha256().newHasher();
+    for (String v : sortedValues) {
+      if (v == null) {
+        hasher.putInt(0);
+      } else {
+        hasher.putString(v, StandardCharsets.UTF_8);
+      }
+    }
+    return new CacheKeyBuilder(DimFilterUtils.IN_CACHE_ID)
+        .appendString(dimension)
+        .appendByte(DimFilterUtils.STRING_SEPARATOR)
+        .appendByteArray(extractionFn == null ? new byte[0] : extractionFn.getCacheKey())
+        .appendByte(DimFilterUtils.STRING_SEPARATOR)
+        .appendByteArray(hasher.hash().asBytes())
+        .build();
   }
 
   private InDimFilter optimizeLookup()
@@ -227,83 +443,32 @@ public class InDimFilter implements DimFilter
     return this;
   }
 
-  @Override
-  public Filter toFilter()
+  private static Iterable<ImmutableBitmap> getBitmapIterable(final Set<String> values, final BitmapIndex bitmapIndex)
   {
-    return new InFilter(
-        dimension,
-        values,
-        longPredicateSupplier,
-        floatPredicateSupplier,
-        doublePredicateSupplier,
-        extractionFn,
-        filterTuning
-    );
+    return Filters.bitmapsFromIndexes(getBitmapIndexIterable(values, bitmapIndex), bitmapIndex);
   }
 
-  @Nullable
-  @Override
-  public RangeSet<String> getDimensionRangeSet(String dimension)
+  private static IntIterable getBitmapIndexIterable(final Set<String> values, final BitmapIndex bitmapIndex)
   {
-    if (!Objects.equals(getDimension(), dimension) || getExtractionFn() != null) {
-      return null;
-    }
-    RangeSet<String> retSet = TreeRangeSet.create();
-    for (String value : values) {
-      String valueEquivalent = NullHandling.nullToEmptyIfNeeded(value);
-      if (valueEquivalent == null) {
-        // Case when SQL compatible null handling is enabled
-        // Range.singleton(null) is invalid, so use the fact that
-        // only null values are less than empty string.
-        retSet.add(Range.lessThan(""));
-      } else {
-        retSet.add(Range.singleton(valueEquivalent));
+    return () -> new IntIterator()
+    {
+      final Iterator<String> iterator = values.iterator();
+
+      @Override
+      public boolean hasNext()
+      {
+        return iterator.hasNext();
       }
-    }
-    return retSet;
+
+      @Override
+      public int nextInt()
+      {
+        return bitmapIndex.getIndex(iterator.next());
+      }
+    };
   }
 
-  @Override
-  public Set<String> getRequiredColumns()
-  {
-    return ImmutableSet.of(dimension);
-  }
-
-  @Override
-  public String toString()
-  {
-    final DimFilterToStringBuilder builder = new DimFilterToStringBuilder();
-    return builder.appendDimension(dimension, extractionFn)
-                  .append(" IN (")
-                  .append(Joiner.on(", ").join(Iterables.transform(values, StringUtils::nullToEmptyNonDruidDataString)))
-                  .append(")")
-                  .appendFilterTuning(filterTuning)
-                  .build();
-  }
-
-  @Override
-  public boolean equals(Object o)
-  {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    InDimFilter that = (InDimFilter) o;
-    return values.equals(that.values) &&
-           dimension.equals(that.dimension) &&
-           Objects.equals(extractionFn, that.extractionFn) &&
-           Objects.equals(filterTuning, that.filterTuning);
-  }
-
-  @Override
-  public int hashCode()
-  {
-    return Objects.hash(values, dimension, extractionFn, filterTuning);
-  }
-
-  private DruidLongPredicate createLongPredicate()
+  private static DruidLongPredicate createLongPredicate(final Set<String> values)
   {
     LongArrayList longs = new LongArrayList(values.size());
     for (String value : values) {
@@ -323,17 +488,7 @@ public class InDimFilter implements DimFilter
     }
   }
 
-  // As the set of filtered values can be large, parsing them as longs should be done only if needed, and only once.
-  // Pass in a common long predicate supplier to all filters created by .toFilter(), so that
-  // we only compute the long hashset/array once per query.
-  // This supplier must be thread-safe, since this DimFilter will be accessed in the query runners.
-  private Supplier<DruidLongPredicate> getLongPredicateSupplier()
-  {
-    Supplier<DruidLongPredicate> longPredicate = this::createLongPredicate;
-    return Suppliers.memoize(longPredicate);
-  }
-
-  private DruidFloatPredicate createFloatPredicate()
+  private static DruidFloatPredicate createFloatPredicate(final Set<String> values)
   {
     IntArrayList floatBits = new IntArrayList(values.size());
     for (String value : values) {
@@ -355,13 +510,7 @@ public class InDimFilter implements DimFilter
     }
   }
 
-  private Supplier<DruidFloatPredicate> getFloatPredicateSupplier()
-  {
-    Supplier<DruidFloatPredicate> floatPredicate = this::createFloatPredicate;
-    return Suppliers.memoize(floatPredicate);
-  }
-
-  private DruidDoublePredicate createDoublePredicate()
+  private static DruidDoublePredicate createDoublePredicate(final Set<String> values)
   {
     LongArrayList doubleBits = new LongArrayList(values.size());
     for (String value : values) {
@@ -383,9 +532,89 @@ public class InDimFilter implements DimFilter
     }
   }
 
-  private Supplier<DruidDoublePredicate> getDoublePredicateSupplier()
+  @VisibleForTesting
+  public static class InFilterDruidPredicateFactory implements DruidPredicateFactory
   {
-    Supplier<DruidDoublePredicate> doublePredicate = this::createDoublePredicate;
-    return Suppliers.memoize(doublePredicate);
+    private final ExtractionFn extractionFn;
+    private final Set<String> values;
+    private final Supplier<DruidLongPredicate> longPredicateSupplier;
+    private final Supplier<DruidFloatPredicate> floatPredicateSupplier;
+    private final Supplier<DruidDoublePredicate> doublePredicateSupplier;
+
+    InFilterDruidPredicateFactory(
+        final ExtractionFn extractionFn,
+        final Set<String> values
+    )
+    {
+      this.extractionFn = extractionFn;
+      this.values = values;
+
+      // As the set of filtered values can be large, parsing them as numbers should be done only if needed, and
+      // only once. Pass in a common long predicate supplier to all filters created by .toFilter(), so that we only
+      // compute the long hashset/array once per query. This supplier must be thread-safe, since this DimFilter will be
+      // accessed in the query runners.
+      this.longPredicateSupplier = Suppliers.memoize(() -> createLongPredicate(values));
+      this.floatPredicateSupplier = Suppliers.memoize(() -> createFloatPredicate(values));
+      this.doublePredicateSupplier = Suppliers.memoize(() -> createDoublePredicate(values));
+    }
+
+    @Override
+    public Predicate<String> makeStringPredicate()
+    {
+      if (extractionFn != null) {
+        return input -> values.contains(extractionFn.apply(input));
+      } else {
+        return values::contains;
+      }
+    }
+
+    @Override
+    public DruidLongPredicate makeLongPredicate()
+    {
+      if (extractionFn != null) {
+        return input -> values.contains(extractionFn.apply(input));
+      } else {
+        return longPredicateSupplier.get();
+      }
+    }
+
+    @Override
+    public DruidFloatPredicate makeFloatPredicate()
+    {
+      if (extractionFn != null) {
+        return input -> values.contains(extractionFn.apply(input));
+      } else {
+        return floatPredicateSupplier.get();
+      }
+    }
+
+    @Override
+    public DruidDoublePredicate makeDoublePredicate()
+    {
+      if (extractionFn != null) {
+        return input -> values.contains(extractionFn.apply(input));
+      }
+      return input -> doublePredicateSupplier.get().applyDouble(input);
+    }
+
+    @Override
+    public boolean equals(Object o)
+    {
+      if (this == o) {
+        return true;
+      }
+      if (o == null || getClass() != o.getClass()) {
+        return false;
+      }
+      InFilterDruidPredicateFactory that = (InFilterDruidPredicateFactory) o;
+      return Objects.equals(extractionFn, that.extractionFn) &&
+             Objects.equals(values, that.values);
+    }
+
+    @Override
+    public int hashCode()
+    {
+      return Objects.hash(extractionFn, values);
+    }
   }
 }
